thesis_export/code/usr2/requirements.txt
pytorch-lightning>=2.0
hydra-core>=1.3
omegaconf>=2.3
opencv-python
torchmetrics
numpy
tqdm
av
mediapipe
